feat(resolver)!: change directory priority for resolution
The package in [nixpkgs][nixpkgs-hyprshade] uses `HYPRSHADE_SHADERS_DIR`
to point to the shader directory instead of
`/usr/share/hyprshade/shaders`, presumably as a workaround for Nix's
isolation preventing the shaders from being installed to the latter. If
this rewrite gets supported in the future, this workaround will likely
still be used.

This commit gives greater priority to user-installed shaders
(`~/.config/{hypr,hyprshade}/shaders`) so that they may override the
ones installed at `HYPRSHADE_SHADERS_DIR`.

[nixpkgs-hyprshade]: https://github.com/NixOS/nixpkgs/blob/24.05/pkgs/applications/window-managers/hyprwm/hyprshade/default.nix